rpma: gpspm: mind the free SQ space
Having pairs of SEND/RECV buffers, one can not assume an order in
which completions from any of them come first to the CQ.
It may happen to collect multiple RECV completions with no
SEND completions in the meantime. But since handling RECV requires
consuming one SQ slot it is required to control is an SQ slot is
available before processing a RECV completion.

The proposed implementation builds up a queue of RECV completions
which are processed only when free SQ slots are available.
The number of free SQ slots determines how many RECV completions
can be processed.